Refactor frontend into modules
