+++ public/app.js
@@ -1,216 +1,57 @@
- (async function(){
-   const container = document.querySelector('.container');
+import { initResize } from './resize.js';
+import TabManager from './tabManager.js';
+import { initDB, getTables, addTable, getRecords, addRecord } from './db.js';
+import { loadDefault, loadView, loadIdentity, bindBack, flipMobile } from './ui.js';
 
-   /*** RESIZE SIDEBAR ***/
-   const handle = document.getElementById('drag-handle');
-   let dragging = false;
-   handle.addEventListener('mousedown', ()=>{ dragging=true; document.body.style.cursor='col-resize'; });
-   document.addEventListener('mousemove', e=>{
-     if(!dragging) return;
-     let w = Math.max(120, Math.min(window.innerWidth-100, e.clientX));
-     container.style.setProperty('--sidebar-w', w+'px');
-   });
-   document.addEventListener('mouseup', ()=>{ dragging=false; document.body.style.cursor=''; });
+(async function(){
+  const container = document.querySelector('.container');
+  initResize();
 
-   /*** TAB MANAGER ***/
-   class TabManager {
-     constructor(tabsEl, panelsContainer){
-       this.tabsEl  = tabsEl;
-       this.panels  = panelsContainer;
-       this.counter = 0;
-       this.meta    = {};
-       tabsEl.addEventListener('click', e=>{
-         if(e.target.matches('.tab')) this.switchTo(e.target.dataset.tab);
-       });
-     }
-     switchTo(id){
-       // highlight
-       this.tabsEl.querySelectorAll('.tab').forEach(t=>
-         t.classList.toggle('active', t.dataset.tab===id)
-       );
-       // show panel
-       this.panels.querySelectorAll('.tab-panel').forEach(p=>
-         p.classList.toggle('active', p.id===id)
-       );
-       // notify
-       this.tabsEl.dispatchEvent(new CustomEvent('tabchange',{detail:id}));
-     }
-     addTab(plural, singular, id){
-       const tabId = id || (this.tabsEl.id + '-custom-' + (++this.counter));
-       // create tab button
-       const btn = document.createElement('button');
-       btn.className = 'tab';
-       btn.textContent = plural;
-       btn.dataset.tab = tabId;
-       this.tabsEl.insertBefore(btn, this.tabsEl.querySelector('.add-tab'));
-       // create panel
-       const p = document.createElement('div');
-       p.className = 'tab-panel';
-       p.id = tabId;
-       this.panels.appendChild(p);
-       // save meta
-       this.meta[tabId] = { plural, singular, panelEl: p };
-       // show it
-       this.switchTo(tabId);
-       return tabId;
-     }
-     getActiveId(){
-       return this.tabsEl.querySelector('.tab.active').dataset.tab;
-     }
-   }
-   // ---- IndexedDB helpers ----
-   let db;
-   const openDB = () => new Promise((res, rej) => {
-     // open existing database without forcing a lower version
-     const req = indexedDB.open('DataViewerDB');
-     req.onsuccess = e => res(e.target.result);
-     req.onerror   = e => {
-       console.error('IndexedDB open error:', e.target.error);
-       rej(e.target.error);
-     };
-   });
-   const getTables = () => new Promise((res, rej) => {
-     const tx = db.transaction('tables','readonly');
-     const store = tx.objectStore('tables');
-     const rq = store.getAll();
-     rq.onsuccess = () => res(rq.result.map(t => ({ tableName: t.tableName, plural: t.plural, singular: t.singular })));  
-     rq.onerror = () => { console.error('IndexedDB getTables error:', rq.error); rej(rq.error); };
-   });
-   const addTable = (plural,singular)=> new Promise((res,rej)=>{
-     // bump DB version to create new object store for this table
-     const newVersion = db.version + 1;
-     db.close();
-     const req = indexedDB.open('DataViewerDB', newVersion);
-     req.onupgradeneeded = e => {
-       const d = e.target.result;
-       // ensure tables store exists
-       if(!d.objectStoreNames.contains('tables')) {
-         d.createObjectStore('tables', {keyPath:'tableName'});
-       }
-       // create data store for this table
-       d.createObjectStore(plural, {keyPath:'id', autoIncrement:true});
-     };
-     req.onsuccess = e => {
-       db = e.target.result;
-       // save metadata
-       const tx = db.transaction('tables','readwrite');
-       const store = tx.objectStore('tables');
-       const rq = store.add({tableName:plural,plural,singular});
-       rq.onsuccess = ()=>res(plural);
-       rq.onerror   = ()=>{ console.error('IndexedDB addTable error:', rq.error); rej(rq.error); };
-     };
-     req.onerror = e => { console.error('IndexedDB addTable error:', e.target.error); rej(e.target.error); };
-   });
-   const getRecords = table=> new Promise((res,rej)=>{
-     const tx = db.transaction(db.objectStoreNames.contains(table) ? table : 'records','readonly');
-     let rq;
-     if (tx.objectStoreNames.contains(table)) {
-       rq = tx.objectStore(table).getAll();
-     } else {
-       const store = tx.objectStore('records');
-       const idx = store.index('tableName');
-       rq = idx.getAll(table);
-     }
-     rq.onsuccess = ()=>res(rq.result);
-     rq.onerror   = ()=>{ console.error('IndexedDB getRecords error:', rq.error); rej(rq.error); };
-   });
-   const addRecord = (table,name)=> new Promise((res,rej)=>{
-     // write to per-table store if exists, otherwise use legacy 'records'
-     const storeName = db.objectStoreNames.contains(table) ? table : 'records';
-     const tx = db.transaction(storeName,'readwrite');
-     const store = tx.objectStore(storeName);
-     const data = storeName === 'records' ? {tableName:table,name} : {name};
-     const rq = store.add(data);
-     rq.onsuccess = ()=>res(rq.result);
-     rq.onerror   = ()=>{ console.error('IndexedDB addRecord error:', rq.error); rej(rq.error); };
-   });
-   
-   // ---- initialize DB and UI ----
-   /*** DATA VIEWER MODULE ***/
-   // instantiate tab manager
-   const dbTabs = new TabManager(
-     document.getElementById('db-tabs'),
-     document.getElementById('db-panels')
-   );
-   const addDataBtn    = document.getElementById('add-data-btn');
-   const addDataTabBtn = document.getElementById('add-data-tab-btn');
+  const dbTabs = new TabManager(
+    document.getElementById('db-tabs'),
+    document.getElementById('db-panels')
+  );
+  const addDataBtn    = document.getElementById('add-data-btn');
+  const addDataTabBtn = document.getElementById('add-data-tab-btn');
 
-   // open DB and load persisted tables
-   db = await openDB();
-   const tables = await getTables();
-   const footer = document.querySelector('.panel.dataviewer .footer');
-   if(tables.length === 0) {
-     footer.style.display = 'none';
-   } else {
-     footer.style.display = 'block';
-   }
-   for(const t of tables){
-     const id = await dbTabs.addTab(t.plural, t.singular, t.tableName);
-     const recs = await getRecords(t.tableName);
-     recs.forEach(r=>{
-       const card = document.createElement('div');
-       card.className = 'card'; card.textContent = r.name;
-       card.addEventListener('click', ()=> loadView(r.name));
-       dbTabs.meta[id].panelEl.appendChild(card);
-     });
-   }
-   if(tables.length>0) dbTabs.switchTo(tables[0].tableName);
+  await initDB();
+  const tables = await getTables();
+  const footer = document.querySelector('.panel.dataviewer .footer');
+  if(tables.length === 0) {
+    footer.style.display = 'none';
+  } else {
+    footer.style.display = 'block';
+  }
+  for(const t of tables){
+    const id = await dbTabs.addTab(t.plural, t.singular, t.tableName);
+    const recs = await getRecords(t.tableName);
+    recs.forEach(r=>{
+      const card = document.createElement('div');
+      card.className = 'card'; card.textContent = r.name;
+      card.addEventListener('click', ()=> loadView(r.name));
+      dbTabs.meta[id].panelEl.appendChild(card);
+    });
+  }
+  if(tables.length>0) dbTabs.switchTo(tables[0].tableName);
 
-   // ---- MAIN CONTENT HELPERS ----
-   const main    = document.getElementById('main');
-   const backBtn = document.getElementById('back-btn');
-   function bindBack(){
-     const b = document.getElementById('back-main');
-     if(!b) return;
-     b.addEventListener('click', ()=>{
-       if(window.innerWidth<=768) container.classList.remove('show-main');
-       else loadDefault();
-     });
-   }
-   function flipMobile(){ if(window.innerWidth<=768) container.classList.add('show-main'); }
-   function loadDefault(){
-     main.innerHTML = `
-       <button class="back" id="back-main">&larr; Back</button>
-       <h2>Main content</h2>
-       <p>Select something from the sidebar…</p>
-     `;
-     bindBack();
-     if(window.innerWidth<=768) container.classList.remove('show-main');
-   }
-   function loadView(title){
-     main.innerHTML = `
-       <button class="back" id="back-main">&larr; Back</button>
-       <h2>${title}</h2>
-       <p>Details about <strong>${title}</strong> …</p>
-     `;
-     bindBack(); flipMobile();
-   }
-   function loadIdentity(){
-     main.innerHTML = `
-       <button class="back" id="back-main">&larr; Back</button>
-       <h2>Identity control</h2>
-       <p>Configure your identity here.</p>
-     `;
-     bindBack(); flipMobile();
-   }
-   document.querySelector('.identity').addEventListener('click', loadIdentity);
-   backBtn.addEventListener('click', ()=>{
-     if(window.innerWidth<=768) container.classList.remove('show-main');
-     else loadDefault();
-   });
-   window.addEventListener('resize', ()=>{ if(window.innerWidth>768) container.classList.remove('show-main'); });
+  const main    = document.getElementById('main');
+  const backBtn = document.getElementById('back-btn');
 
-  // update footer text on tab switch
+  document.querySelector('.identity').addEventListener('click', loadIdentity);
+  backBtn.addEventListener('click', ()=>{
+    if(window.innerWidth<=768) container.classList.remove('show-main');
+    else loadDefault();
+  });
+  window.addEventListener('resize', ()=>{ if(window.innerWidth>768) container.classList.remove('show-main'); });
+
   dbTabs.tabsEl.addEventListener('tabchange', e=>{
     const m = dbTabs.meta[e.detail];
     addDataBtn.textContent = `+ New ${m.singular}`;
   });
-  // set initial footer button label if there are tables
   if (tables.length > 0) {
     const m = dbTabs.meta[dbTabs.getActiveId()];
     addDataBtn.textContent = `+ New ${m.singular}`;
   }
-  // clicking "+ New ..." loads card form
   addDataBtn.addEventListener('click', async ()=>{
      const id = dbTabs.getActiveId();
      const { singular, panelEl } = dbTabs.meta[id];
@@ -236,32 +77,29 @@
        loadDefault();
      });
      bindBack(); flipMobile();
-   });
-   // clicking "+" in tab bar loads new-tab form
-   addDataTabBtn.addEventListener('click', ()=>{
-     main.innerHTML = `
-       <button class="back" id="back-main">&larr; Back</button>
-       <h2>New Data Tab</h2>
-       <form id="tab-form">
-         <label>Tab Name (plural): <input type="text" id="tab-name" required></label><br><br>
-         <label>Item Name (singular): <input type="text" id="item-name" required></label><br><br>
-         <button>Create Tab</button>
-       </form>
-     `;
-     document.getElementById('tab-form').addEventListener('submit', async e=>{
-       e.preventDefault();
-       const p = document.getElementById('tab-name').value.trim();
-       const s = document.getElementById('item-name').value.trim();
-        if(p && s){
-          const tName = await addTable(p, s);
-          await dbTabs.addTab(p, s, tName);
-          // show footer when first table is created
-          footer.style.display = 'block';
-          addDataBtn.textContent = `+ New ${s}`;
-        }
-       loadDefault();
-     });
-     bindBack(); flipMobile();
-   });
-  // end init
- })();
+  });
+  addDataTabBtn.addEventListener('click', ()=>{
+    main.innerHTML = `
+      <button class="back" id="back-main">&larr; Back</button>
+      <h2>New Data Tab</h2>
+      <form id="tab-form">
+        <label>Tab Name (plural): <input type="text" id="tab-name" required></label><br><br>
+        <label>Item Name (singular): <input type="text" id="item-name" required></label><br><br>
+        <button>Create Tab</button>
+      </form>
+    `;
+    document.getElementById('tab-form').addEventListener('submit', async e=>{
+      e.preventDefault();
+      const p = document.getElementById('tab-name').value.trim();
+      const s = document.getElementById('item-name').value.trim();
+       if(p && s){
+         const tName = await addTable(p, s);
+         await dbTabs.addTab(p, s, tName);
+         footer.style.display = 'block';
+         addDataBtn.textContent = `+ New ${s}`;
+       }
+      loadDefault();
+    });
+    bindBack(); flipMobile();
+  });
+})();
