+++ public/index.html
@@ -36,6 +36,6 @@
   </div>
 </div>
 
-  <script src="app.js"></script>
+  <script type="module" src="app.js"></script>
 </body>
 </html>
